update README with control model and monitor model
The project is going to have executables as control services. They are
programmed with two models; control model and monitor model. The former
is designed to dispatch ALSA control event from userspace applications,
and convert it according to model-specific protocols, then operate units
in node on IEEE 1394 bus. The latter is designed to dispatch events from
units in node on IEEE 1394 bus, and convert it to value for ALSA control
element, then operate the control element.

This commit adds illustration about the two models.

Signed-off-by: Takashi Sakamoto <o-takashi@sakamocchi.jp>