dispatcher: add dispatcher structure to run glib::MainContext and glib::MainLoop in own thread
This project produces executables for services to control unit according
to events from userspace application.

This commit adds a dispatcher structure to dispatch the event. The
dispatcher utilizes glib::MainContext and glib::MainLoop running in
own thread. When dropping the structure, glib::MainContext is canceled
and exit the thread. The main thread joins the exit safely.

Signed-off-by: Takashi Sakamoto <o-takashi@sakamocchi.jp>