tascam: async_unit: handle surface event to emit ALSA Sequencer event
This commit adds support for toggle buttons to emit ALSA sequencer event.
The release/push events of button are dispatched and converted to ALSA
sequencer event in which channel equals to 0, param equals to the position
of button, and value equals to 0 (=release) or 127 (=push). The map betwen
position and button is created dynamically.

Signed-off-by: Takashi Sakamoto <o-takashi@sakamocchi.jp>