dg00x: monitor_ctl: handle control event for monitor features
This commit adds monitor controls for activation and input gains. I note
that actual functionality depends on the state of packet streaming. When
packet streaming runs, everything works well. Without packet streaming,
monitor configuration is not updated according to quadlet write transaction
even if the transaction actually changes the value of registers. For this
quirk, the registers are updated according to cached state whenever packet
streaming is detected.

Signed-off-by: Takashi Sakamoto <o-takashi@sakamocchi.jp>