tascam: async_unit: handle surface event for fader
FE-8 has 8 faders with touch sensors. The sensors performs like buttons
in a point of control surface event.

This commit dispatches the control surface event for sensors. When sensors
are released, the value of fader is emitted as ALSA sequencer event with
control data in which channel equals to 0, params equals to the position
of fader, and value equals to the value of fader.

Signed-off-by: Takashi Sakamoto <o-takashi@sakamocchi.jp>