tascam: isoc_console_unit/isoc_rack_unit: support ALSA Control event
The aim of service program is to dispatch ALSA control event from
applications, then convert the event to model-specific message and
transfer it to the unit.

This commit adds support of the routine. ALSA control character device
is opened and closures are assigned to GSource created for the character
device.

Signed-off-by: Takashi Sakamoto <o-takashi@sakamocchi.jp>