tascam: isoc_console_unit: handle surface event to emit ALSA Sequencer event
FW-1884 and FW-1082 have some simple buttons and dials, in addition to
toggle buttons, bank buttons and transport buttons.

This commit adds support for these buttons to emit ALSA sequencer event.
The release/push events of button are dispatched and converted to ALSA
sequencer event in which channel equals to 0, param equals to the position
of button, and value equals to 0 (=release) or 127 (=push). The map betwen
position and button is created dynamically.

Signed-off-by: Takashi Sakamoto <o-takashi@sakamocchi.jp>