tascam: isoc_console_unit: add support for ALSA Sequencer
The model of console type has physical control elements such as faders,
buttons, dial, and knob in its surface. These are familiar to MIDI
messages rather than control events.

This commit adds support for ALSA sequencer to console models. A client
is registered when service program runs. A port is added by the client and
is going to dispatch ALSA sequencer event. The event is expected to be
control type and includes control data.

Signed-off-by: Takashi Sakamoto <o-takashi@sakamocchi.jp>