tascam: async_unit: handle ALSA Sequencer application event to turn LED on/off
FE-8 have several LEDs as a feature of control surface. The LEDs are
controllable by software.

This commit dispatches ALSA sequencer events to bright the LEDs. Control
type of ALSA sequencer event is used for the purpose. Userspace
applications transfer the event in which channel equalds to 0, param equals
to the position of LED, and value equals to 0 (=off) or 127 (=on).

Signed-off-by: Takashi Sakamoto <o-takashi@sakamocchi.jp>