efw: add support for commands in port configuration category
In vendor-specific transaction for Fireworks board module, commands are
categorized according to their purposes.

This commit adds support for commands in port configuration category.
When receiving the commands, Fireworks board module configures mirroring
for control room, the mode of digital interface, phantom powering, and
the map of playback/capture to physical output/input.

Signed-off-by: Takashi Sakamoto <o-takashi@sakamocchi.jp>