efw: fix failure to configure sampling rate
In transaction of Fireworks board module, sampling rate is configured
by literal number. On the other hand, the handler for sampling-rate
control element attempts to set index number in model-specific table
for supported sampling rate. As a result, invalid number is going to
be configured.

This commit fixes failure to configure sampling rate.

Fixes: 4b43427b9ebd ("efw: add clock control elements")
Signed-off-by: Takashi Sakamoto <o-takashi@sakamocchi.jp>