efw: fix failure to configure clock source
In transaction of Fireworks board module, clock source is configured by
magic number. On the other hand, the handler for clock-source control
element attempts to set index number in model-specific table for
supported clock source. As a result, invalid number is going to be
configured.

This commit fixes failure to configure clock source.

Fixes: 4b43427b9ebd ("efw: add clock control elements")
Signed-off-by: Takashi Sakamoto <o-takashi@sakamocchi.jp>