efw: cache current states of nominal input level for FPGA model
In FPGA models, the response to read request for nominal input
level is the same even if any change is actually applied. The response
always includes the same value.

This commit is for the quirk. The cache is added for current status of
nominal input level for FPGA model. In the case of FPGA models, nominal
input level is configured to be the default value so that the cache works
properly.

Signed-off-by: Takashi Sakamoto <o-takashi@sakamocchi.jp>