ta1394: add enumerations for command type of AV/C transaction
AV/C command protocol is one of specialization of FCP in IEC 61883-1.
The first byte of transaction frame for command has 0b0000 in
Command/transaction set (CTS) encoding field. The rest is used for
command type.

This commit adds enumerations and a pair of parser and builder for the
commands.

Signed-off-by: Takashi Sakamoto <o-takashi@sakamocchi.jp>