bebob: add unit structure and service program
BridgeCo AG designed the series of ASICs for audio applications on IEEE
1394 bus. DM1000/DM1100/DM1500 ASICs are actually used for many vendors.
The ASIC integrate ARM core and work with downloaded firmware so that
the vendors can optimize and customize its behaviour.

ALSA bebob driver supports the ASICs in a point of packet streaming. On
the other hand, no application had been developed for any other
functionalities.

This commit adds an executable, snd-bebob-ctl-service, to dispatch
events from ALSA control applications to control the functionalities.

Signed-off-by: Takashi Sakamoto <o-takashi@sakamocchi.jp>