services: add crate to produce executables as services
The main purpose of this project is to develop service programs to
control audio and music unit on IEEE 1394 bus supported by ALSA firewire
stack.

This commit adds a crate as 'services' and push codes for the programs
into it. It's not applicable for standard directory layout of binary
project since including multiple programs, therefore [[bin]] directive
is used.

Signed-off-by: Takashi Sakamoto <o-takashi@sakamocchi.jp>