use workspace to aggregate crates
At present, all of codes in this project are under the same crate;
`snd-fw-ctl-services`. This forces to rebuild whole codes including
executables when the part of code is changed. This increases build
time so long by several ten seconds. It's not preferable.

This commit is a preparation to split the codes into several crates
to shorten the build time. As first place, this commit uses workspace
to aggregate crates added for future.

Signed-off-by: Takashi Sakamoto <o-takashi@sakamocchi.jp>