alsa-ctl-tlv-codec: add crate to encode/decode TLV data in ALSA control interface
In ALSA control interface, control element set can have data in TLV
(Type-Length-Value) style. This data is added by kernel drivers or
userspace applications, then it's referred by the other userspace
applications.

This commit adds a crate, alsa-ctl-tlv-codec to parse and build the
data of TLV style. The crate is purely written by Rust and no dependency
to the other crate or software.

Signed-off-by: Takashi Sakamoto <o-takashi@sakamocchi.jp>