core: use slice of u32 instead of i32 for data of TLV
This commit is a preparation to use alsa-ctl-tlv-codec crate to build
data of TLV. The alsactl crate has a bug to accept array of i32 element,
thus current code include structures to have static array of i32 element.
They should be u32 array to use the crate. Instead, methods of CardCntr
structure are written to convert these arrays.

Signed-off-by: Takashi Sakamoto <o-takashi@sakamocchi.jp>