ieee1212-config-rom: code refactoring to use slice instead of new string for text of leaf entry
The crate uses std::str::from_utf8() to convert from &[u8] to &str, then
generate new String. The resultant slice has the same lifetime of the
source. For many cases, the owner of source uses the method therefore
it's possible to return the resultant slice since it should be owned by
the owner.

This commit uses string slice instead of new String to avoid memory
allocation.

Signed-off-by: Takashi Sakamoto <o-takashi@sakamocchi.jp>