ieee1212-config-rom: code refactoring to avoid data duplication
The leaf entry of Configuration ROM includes a string if it includes
textual descriptor. Current implementation uses Vec<u8> to duplicate
the raw data of string. However it's possible to refer to byte slice
instead since the owner of raw data for Configuration ROM is the user
of parse result.

This commit uses byte slice for data of leaf entry to avoid memory
allocation.

Signed-off-by: Takashi Sakamoto <o-takashi@sakamocchi.jp>