ieee1212-config-rom: add structure to represent content of configuration ROM
Current implementation returns Vec<Entry> as a result of parsing
content of Configuration ROM. However, it's convenient to access to bus
information for applications.

This commit adds ConfigRom structure for the purpose. It implements
std::convert::TryFrom trait to parse &[u8] as raw data of Configuration
ROM. The conversion can fail with ConfigRomParseError to deliver
information of error cause.

Signed-off-by: Takashi Sakamoto <o-takashi@sakamocchi.jp>