dg00x: use helper function in ta1394 to pick up model information
The content of Configuration ROM in Digidesign Digi00x series is not
compliant to 1394TA specification in a point of specifier ID of
descriptor for vendor leaf entry. However, newly added desc module
in ieee1212 crate can parse it.

This commit uses helper function in ta1394 crate to parse it.

Signed-off-by: Takashi Sakamoto <o-takashi@sakamocchi.jp>