dice-protocols: tcat: extension: add trait and its implementation for command feature of TCAT protocol extension
In TCAT protocol extension, some commands are supported to operate
memory space on the target unit. The commands allow controller node to
get and set current configuration to the unit as well as store/load
the configurations to/from flash memory.

This commit adds trait and its implementation for the section.

Signed-off-by: Takashi Sakamoto <o-takashi@sakamocchi.jp>