dice-protocols: tcat: add module for protocol defined by TCAT
The operation for ASICs of DICE is performed by asynchronous transaction
to private space of CSR in target unit. The space is separated to
several sections for each functionality. In general protocol, four
sections are used.

This commit adds module for trait and its implementation of the general
protocol. For asynchronous transaction, libhinawa and its Rust binding are used.

Signed-off-by: Takashi Sakamoto <o-takashi@sakamocchi.jp>