dice-protocols: tcat: extension: add local trait and its implementation for stream format section of TCAT protocol extension
In TCAT protocol extension, format of packet transferred and received by
the target unit is represented in stream format section. The section
includes information about the format according to the result of
operation in cmd section. The controller node can get and set format of
packet, then refrects the change into the target unit by operation of
cmd section.

This commit adds trait and its implementation of the section.

Signed-off-by: Takashi Sakamoto <o-takashi@sakamocchi.jp>