dice-runtime: add trait for specification of TCD2210/TCD2220
The enhanced features by TCAT protocol extension are available
TCD2210/TCD2220 ASICs of DICE. These ASICs have common specification for
the feature.

This commit adds a trait to represent the feature. Most vendors
customized the ASICs to reduce functionality. The trait has some
associated constants so that implementation of the trait can represent
the customization.

Signed-off-by: Takashi Sakamoto <o-takashi@sakamocchi.jp>