dice-protocols: maudio: add methods specific to ProFire 2626
M-Audio ProFire 2626 supports two operations in application section
of TCAT protocol extension. It has second optical interface and allows
software to decide which types of signal is used for the interface.
Additionally, it supports two types of mode for standalone converter.

This commit adds the above operations to protocol implementation.

Signed-off-by: Takashi Sakamoto <o-takashi@sakamocchi.jp>