dice-protocols: tcelectronic: shell: add common structure for mixer
The 'shell' type of models in Konnekt series commonly supports internal
multiplexer, which multiplexes inputs from analog, digital, and stream
to a pair of output with configured gain. The configuration of multiplexer
locates in a segment including model-specific functionalities.

Some models supports reverb effect and allows software to configure the
effect in the same segment.

The multiplexer have metering. The models have a segment for it.

This commit adds structures for them.

Signed-off-by: Takashi Sakamoto <o-takashi@sakamocchi.jp>