dice-protocols: tcelectronic: add structure and trait for segment in protocol for Konnekt series
The memory space of models in Konnekt series consists of several
segments which has sub-effect to control hardware by write transaction.
In the segment, quadlet-aligned data represents reach configuration.
Current configuration is available by read transaction to the segment.

This commit adds generic structure to represent the segment as well as
trait implementation to read/write content of the segment.

Signed-off-by: Takashi Sakamoto <o-takashi@sakamocchi.jp>