dice-protocols: tcelectronic: add trait implementation for mixer segment of Studio Konnekt 48
Studio Konnekt 48 has internal mixer to multiplex signals from analog,
digital, stream inputs to three pairs of output with gains. The mixer
has 12 slots for the pair of input. The first 5 slots are fixed to
each input, and the others are available for user's configuration.
The mixer also has meter.

Two segments are used to represent state and meter of mixer. This commit
implements the segments.

Signed-off-by: Takashi Sakamoto <o-takashi@sakamocchi.jp>