dice-protocols: tcelectronic: add structure for data of channel strip effect
Some models in Konnekt series support channel strip effect including
equalizer, compressor, limitter, and deesser. The effect is called as
'Fabrik C'. The segment for the effect consists of 144 byte data for
state of each channel and 28 byte data for meter of each channel.
The sequence of channel includes 4 byte padding between two pair of
channel.

This commit adds structure and builder/parser for the state and meter
of effect.

Signed-off-by: Takashi Sakamoto <o-takashi@sakamocchi.jp>