services: parse open error and print friendly
All of the service programs access to character devices of Linux sound
subsystem and Linux FireWire subsystem, therefore effective user and
group of process should have access permission to them. Typical failure
to run the service program comes from it.

This commit attempts to detect the failure and print message friendly
to users for the case.

Signed-off-by: Takashi Sakamoto <o-takashi@sakamocchi.jp>