ff-protocols: former: add trait implementation for meter protocol of Fireface 800
In RME Fireface series, hardware metering is available by read block
transaction. The read data consists of some octlets and quadlets. The
octlet meter seems to be for the result to processing audio signal
in DSP and the quadlet meter seems to be for direct hardware inputs and
outputs.

This commit adds trait implementation for the meter protocol.

Signed-off-by: Takashi Sakamoto <o-takashi@sakamocchi.jp>