core: dispatcher: fix thread blocking in shutdown process due to full queue
When implementation of RuntimeOperation::listen() fails due to ENOMEM
error in call of ioctl(2) with SNDRV_CTL_IOCTL_ELEM_ADD, the process
should be going to shutdown. However, preceeding calls already queue
some ALSA control element event in kernel space. Running dispatcher
reads the events and sends to std::sync::mpsc, then possible to reach
the limit of queue in userspace. As a result, the sender thread is
blocked due to the lack of rest in the queue. In this case, the
shutdown corrupts due to thread joining even if glib::MainLoop is set
to quit.

Current implementation gives an internal API for the combination of
glib::MainLoop::quit() and std::Thread::join(). However, to release
the blocked thread, it's required to hollow the queue. This should
be done after scheduling stop of dispatchers in sender side since no
additional events should not be queued.

This commit fixes the bug. The commit adds an independent internal
API to schedule stop of dispatcher. After calling the API, runtime
instance is expected to hollow the queue, then drop dispatcher to
join thread.

I note that Linux kernel v5.13 or later extends the capacity of
user-defined element set. the ioctl error will becomes rare.

Signed-off-by: Takashi Sakamoto <o-takashi@sakamocchi.jp>