efw-protocols: add trait for protocol
Near future, the operation related to on-board flash memory will be
implemented in protocols crate. For the case, the implementation
should work without any assist of ALSA fireworks driver. On the
other hand, current implementation for the rest of protocol heavily
depends on the driver.

This commit adds a trait as wrapper as quick solution.

Signed-off-by: Takashi Sakamoto <o-takashi@sakamocchi.jp>