dice-protocols: focusrite: add protocol implementation of DSP effect in Pro 24 DSP
This commit adds supoprt for compressor and reverb effects, but no support
for equalizer effect. For equalizer, each coefficient block includes
5 quadlets data for each of 4 frequency bands. All of 5 quadlets are
changed even when one of equalizer parameters (center frequency, gain,
quality factor, and shelving) is operated in GUI. It's hard to figure
out the formula between parameter operation and changes of 5 quadlets.

Additionally, the source/destination ID of channel strip effect input
and output is different between low and middle sampling rate, as well
as the source/destination ID of reverb effect. The commit adds entries
corresponding to the low sampling rate. It's TODO to support entries
for middle sampling rate.

Signed-off-by: Takashi Sakamoto <o-takashi@sakamocchi.jp>