runtime/core: fix build failure for dice runtime
For debug logging for enumerated type of ALSA control element, the generic
function to add the element requires labels to have std::fmt::Debug, while
dice runtime implementation has missing annotation in helper function
which calls the function. This brings FTBFS.

This commit rearranges the function. Fortunately the function already
generates vector of str from the labels and it is available for debugging.

Fixes: 8627bab68421 ("runtime/core: add tracing events for operations related to ALSA control interface")
Signed-off-by: Takashi Sakamoto <o-takashi@sakamocchi.jp>