protocols/dice: tcat: add traits and their implementations of parameters in application section
Hardware vendors implement unique functions in application section of
TCAT protocol extension. The parameters of functions is expressed in
regions of the section. Any change of parameter is also expressed in
change of region. Additionally, notification of DICE general protocol
can includes message relevant to the region.

This commit adds traits and their implementation for the section.

Signed-off-by: Takashi Sakamoto <o-takashi@sakamocchi.jp>