protocols/dice: tcat: fulfill serializer of global parameters in TCAT general protocol
For future work, A pair of serializer and deserializer is required for
global parameters in TCAT general protocol, while current implementation
has serializer with less functions.

This commit fulfill the serializer. Fortunately, actual devices can returns
invalid content of address space for the parameters and current
implementation corrects it with hard-coded values, thus the content is not
necessarily recoverable. This commit adds the serializer by the fashion of
best effort.

Signed-off-by: Takashi Sakamoto <o-takashi@sakamocchi.jp>