protocols/fireface: fix invalid index of microphone inputs
The commands for microphone inputs (phantom powering and instrument mode)
are not effective since the index is invalid. This commit fixes the bug
as well as wrong channel count in control element.

Additionally, the settings for instrument is prior to the setting
for phantom powering. This commit supports it.

Reference: https://github.com/alsa-project/snd-firewire-ctl-services/issues/44
Signed-off-by: Takashi Sakamoto <o-takashi@sakamocchi.jp>