runtime/core: use feature of nix crate for UNIX signals
The nix crate introduced features in version 0.24.0 release to enable
and disable functions, like signals. In the timing, all of the features
were enabled as a default. However, in version 0.27.0 release, the
features are disabled as a default. Applications should indicate these
features to enable used functions.

It causes FTBFS reported in repository in github.com. This commit fix the
issue to enable features required to signals.

Reference: https://github.com/alsa-project/snd-firewire-ctl-services/issues/176
Signed-off-by: Takashi Sakamoto <o-takashi@sakamocchi.jp>