protocols/dice: fix to decode the sequence of labels when followed by garbage bytes
An issue #200 filed at github.com reports a case that the current
implementation of decoder for the sequence of labels does not work
well.

https://github.com/alsa-project/snd-firewire-ctl-services/issues/200

When working with a test including the retrieved byte dump, it appears that
the implementation attempts to decode the byte data after detecting a pair
of double backslashes (= terminator of the sequence), then face the
malformed UTF-8 byte sequence.

This commit fixes the implementation so that it cancels iteration when
finding an empty element of array equivalent to the double backslashes.
The fix resolves the test.

Signed-off-by: Takashi Sakamoto <o-takashi@sakamocchi.jp>